workflow engine: rename db1 lifecycle header to avoid db2/lifecycle.h collision
The W2 accessor header was named db1/lifecycle.h, whose basename collides with
the existing db2/lifecycle.h that kb_main.c (and ~10 other files) include as
"lifecycle.h"; with -Idb1 ahead of -Idb2 it shadowed the db2 header and broke
the aimee-kb build (implicit db2_init/db2_health_probe/...). Renamed to
db1/wfe_store.{c,h} (unique basename) and updated all workflow-engine includers.
No behavior change; client + server + kb all build clean.
