+++ src/tests/test_wfe_autonomy.c
@@ -6,7 +6,7 @@
 #include <sys/stat.h>
 
 #include "db1.h"
-#include "lifecycle.h"
+#include "wfe_store.h"
 #include "wfe_approval.h"
 #include "wfe_autonomy.h"
 #include "wfe_engine.h"
