+++ src/tests/test_wfe_engine.c
@@ -8,7 +8,7 @@
 #include <sys/stat.h>
 
 #include "db1.h"
-#include "lifecycle.h"
+#include "wfe_store.h"
 #include "wfe_engine.h"
 
 /* mini workflow: draft -> gate -> pr -> done(merge); gate loops to draft. */
