+++ src/workflow/wfe_approval.c
@@ -8,7 +8,7 @@
 
 #include "aimee_home.h"
 #include "cJSON.h"
-#include "lifecycle.h"
+#include "wfe_store.h"
 #include "wfe_def.h"
 #include "wfe_engine.h"
 #include "wfe_iface.h"
