+++ src/workflow/wfe_autonomy.c
@@ -5,7 +5,7 @@
 #include <string.h>
 
 #include "cJSON.h"
-#include "lifecycle.h"
+#include "wfe_store.h"
 #include "wfe_approval.h"
 #include "wfe_def.h"
 #include "wfe_engine.h"
